#1 issue solved by creating several function which check whether Tomcat service has been installed already. If positive the service will be removed and reinstalled. Check if port 8080 is in LISTEN state to avoid port conflict

+++ _output/tomcatzip/8.0.8/tools/chocolateyInstall.ps1
@@ -2,15 +2,61 @@
 $version = '8'
 $serviceName = "tomcat${version}"
 $extractionPath = "C:/Program Files/$packageName"
+$catalinaHome = "${extractionPath}/apache-tomcat-8.0.8"
+$port = '8080'
 
-Install-ChocolateyZipPackage "$packageName" "http://apache.hippo.nl/tomcat/tomcat-8/v8.0.8/bin/apache-tomcat-8.0.8-windows-x86.zip" "$extractionPath" "http://apache.mirrors.hostinginnederland.nl/tomcat/tomcat-8/v8.0.8/bin/apache-tomcat-8.0.8-windows-x64.zip"
+function installService() {
+    if (serviceExists) {
+        Write-Host "Apache Tomcat service will be installed"
+		$createService = "${catalinaHome}/bin/service.bat"
+		& $createService install $serviceName
+    }
+}
+
+function startService() {
+    if (portAvailable) {
+        Write-Host "Apache Tomcat service will be started"
+		start-service $serviceName
+    } else {
+		Write-ChocolateyFailure 'Apache Tomcat' "port $port is unavailable"
+		throw 
+    }
+}
+
+function portAvailable() {
+	Try {
+		$connection = (New-Object Net.Sockets.TcpClient)
+		$connection.Connect("127.0.0.1",$port)
+		return $FALSE
+	} Catch {
+		return $TRUE
+	}
+}
+
+function serviceExists() {
+	$service = Get-Service -Name $serviceName -ErrorAction SilentlyContinue
+	if ($service.Length -gt 0) {
+        Write-Host "$serviceName service already exists and will be removed"
+		stop-service $serviceName
+		$service = Get-WmiObject -Class Win32_Service -Filter "Name='$serviceName'"
+		$service.delete()
+		return $TRUE
+    } else {
+		return $TRUE
+    }
+}
+
+Install-ChocolateyZipPackage "$packageName" "http://apache.hippo.nl/tomcat/tomcat-8/v8.0.8/bin/apache-tomcat-8.0.8-windows-x86.zip" "$extractionPath" "http://mirrors.supportex.net/apache/tomcat/tomcat-8/v8.0.8/bin/apache-tomcat-8.0.8-windows-x64.zip"
+
+Write-Host "Adding Apache Tomcat 8.0.8 executables to Chocolatey's bin path"
 Get-ChocolateyBins "$extractionPath"
 
-$catalinaHome = "${extractionPath}/apache-tomcat-8.0.8"
-
+Write-Host "Setting Environment Variable CATALINA_HOME to $catalinaHome"
 Set-Variable CATALINA_HOME $catalinaHome
-$Env:CATALINA_HOME += $catalinaHome
+Install-ChocolateyEnvironmentVariable 'CATALINA_HOME' $catalinaHome 'Machine'
 
-$createService = "${catalinaHome}/bin/service.bat"
-& $createService install $serviceName 
-net start $serviceName
+installService
+
+startService
+
+Write-ChocolateySuccess 'Apache Tomcat 8.0.8'
